Added material design and worksheet crud to db

+++ ConstructionDiary/ViewModels/WorkSheet/MaterialsVM.cs
@@ -8,7 +8,7 @@ namespace ConstructionDiary.ViewModels.WorkSheet
     public class MaterialsVM
     {
         public int id { get; set; }
-        public decimal amount { get; set; }
+        public double amount { get; set; }
         public string name { get; set; }
 
     }
